pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tyr"
version = "0.1.0"
description = "Polymarket arbitrage bot"
requires-python = ">=3.12"
dependencies = [
    "py-clob-client>=0.34.6",
    "httpx>=0.27.0",
    "websockets>=13.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "rapidfuzz>=3.9.0",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "rich>=13.9.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[project.scripts]
tyr = "tyr.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["tyr*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
